Use correct types, fix a lost of warnings, and reformat a large part
This includes a lot of changes, mainly in the three areas:
* Correct types (NSInteger-ization, string type typedefs, enums, etc.)
* Fix various warnings
* Manual reformatting of code to be more readable

I'm sorry I could not split this huge commit into neat separate commits.

+++ AppKit/NSBitmapImageRep.m
@@ -497,7 +497,7 @@ NSBitmapImageRepPropertyKey NSImageCurrentFrame = @"NSImageCurrentFrame";
 
    NSInteger i,numberOfComponents=[color numberOfComponents];
    CGFloat   components[numberOfComponents];
-   unsigned  pixels[numberOfComponents];
+   NSUInteger  pixels[numberOfComponents];
 
 	// Extract the RGBA components
    [color getComponents:components];
